requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
pydantic-settings>=2.0.0

# ML Libraries
torch==2.1.0
transformers==4.35.0
sentencepiece==0.1.99
protobuf==4.24.4

# PDF Processing
pymupdf==1.23.4  # PyMuPDF
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# Text Processing
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2